collapse implementation first pass.
